Measure xmhf-runtime before xmhf_sl_handle_rt_rela_dyn
