+++ xmhf/src/xmhf-core/xmhf-runtime/runtime.lds.S
@@ -66,7 +66,7 @@ ENTRY(xmhf_runtime_entry)
 
 MEMORY
 {
-  all (rwxai) : ORIGIN = __TARGET_BASE, LENGTH = 512M /* length is arbitrary */
+  all (rwxai) : ORIGIN = __TARGET_BASE, LENGTH = 1024M /* length is arbitrary */
   debug (rwxai) : ORIGIN = 0, LENGTH = 1024M
   unaccounted (rwxai) : ORIGIN = 0, LENGTH = 0 /* see section .unaccounted at end */
 }
