+++ xmhf/src/xmhf-core/xmhf-secureloader/arch/x86/sl-x86.c
@@ -55,10 +55,6 @@
 // Implement SRTM and the chain of measurement missed in DRTM.
 
 #include <xmhf.h>
-#include "../../hash/hash.h"
-
-#define TPM_PCR_BOOT_STATE   (7)
-#define TPM_PCR_DRTM_IMAGE  (17)
 
 // 1 pages of memory in untrusted SL memory for Intel to disable all DMA accesses
 // According to implementation, they should be simply memset to 0
@@ -337,56 +333,6 @@ void xmhf_sl_arch_early_dmaprot_init(u32 runtime_size)
 }
 #endif /* __DMAP__ */
 
-/// @brief Return true iff xmhf-sl can use the physical TPM device (either TPM 1.2 or TPM 2.0).
-/// @param out_tpm 
-/// @param out_tpm_fp 
-/// @return 
-static bool _is_tpm_present(struct tpm_if **out_tpm, struct tpm_if_fp **out_tpm_fp)
-{
-    struct tpm_if *tpm = get_tpm();
-    struct tpm_if_fp *tpm_fp = NULL;
-
-// [TODO][Github-XMHF64 Issue 13] A QEMU issue? If XMHF accesses SW TPM of QEMU with tpm20.c, then QEMU reports the 
-// error "Buffer Too Small" when loading the OS bootloader
-#if defined(__DEBUG_QEMU__) && defined(__UEFI__) && !defined(__FORCE_TPM_1_2__)
-    {
-        printf("xmhf-sl: No support of SW TPM2.0 in QEMU!\n");
-        return false;
-    }
-#endif // defined(__DEBUG_QEMU__) && defined(__UEFI__)
-
-    if(!tpm)
-    {
-        printf("xmhf-sl: Failed to get <tpm>!\n");
-        return false;
-    }
-
-    if(!tpm_detect())
-    {
-        printf("xmhf-sl: Failed to get TPM version!\n");
-        return false;
-    }
-
-    tpm_fp = (struct tpm_if_fp *)get_tpm_fp();
-    if(!tpm_fp)
-    {
-        printf("xmhf-sl: Failed to get <tpm_fp>!\n");
-        return false;
-    }
-
-    // Check TPM versions
-    if((tpm->major != TPM12_VER_MAJOR) && (tpm->major != TPM20_VER_MAJOR))
-    {
-        printf("xmhf-sl: Unknown TPM version!\n");
-        return false;
-    }
-
-    // On success
-    *out_tpm = tpm;
-    *out_tpm_fp = tpm_fp;
-    return true;
-}
-
 #ifdef __AMD64__
 static void _invoke_runtime_entrypoint(RPB *rpb, u64 ptba)
 #elif defined(__I386__)
@@ -404,16 +350,6 @@ static void _invoke_runtime_entrypoint(RPB *rpb, u32 ptba)
 #endif /* !defined(__I386__) && !defined(__AMD64__) */
 }
 
-static hva_t xmhf_runtime_get_base_hvaddr(RPB *rpb)
-{
-#ifdef __XMHF_PIE_RUNTIME__
-    return (hva_t)(__TARGET_BASE + rpb->XtVmmRelocationOffset);
-#else /* !__XMHF_PIE_RUNTIME__ */
-    (void)rpb;
-    return (hva_t)__TARGET_BASE;
-#endif /* __XMHF_PIE_RUNTIME__ */
-}
-
 void xmhf_sl_arch_xfer_control_to_runtime(RPB *rpb)
 {
 #ifdef __AMD64__
@@ -427,10 +363,6 @@ void xmhf_sl_arch_xfer_control_to_runtime(RPB *rpb)
 	TSSENTRY *t;
 	hva_t tss_base;
 	hva_t gdt_base;
-
-    struct tpm_if *tpm = NULL;
-    struct tpm_if_fp *tpm_fp = NULL;
-    bool found_tpm = false;
 
 	#ifndef __XMHF_VERIFICATION__
 		//setup runtime TSS
@@ -470,102 +402,6 @@ void xmhf_sl_arch_xfer_control_to_runtime(RPB *rpb)
 	#endif
 
 	printf("SL: setup runtime paging structures.\n");
-
-    found_tpm = _is_tpm_present(&tpm, &tpm_fp);
-    if(!found_tpm)
-    {
-        // XMHF cannot use TPM. Warn it loud.
-        printf("**********************************************************************************************\n");
-        printf("[INSECURITY] XMHF CANNOT USE TPM!\n");
-        printf("**********************************************************************************************\n");
-        _invoke_runtime_entrypoint(rpb, ptba);
-    }
-
-    // Measure xmhf-runtime into TPM PCRs
-    // [NOTE] Even with DRTM enabled, xmhf-bootloader must measure xmhf-runtime into PCR7 to maintain the security of 
-    // red OS. Otherwise, remote attackers can compromise xmhf-runtime to steal Bitlocker "volume master key" and reboot
-    // immediately. So the attacker can get the secret without getting exposed in PCR7 (or anywhere in PCR0-15). 
-    {
-        union sha_digest digest = {0};
-        int result = 0;
-        hva_t xmhf_rt_base = xmhf_runtime_get_base_hvaddr(rpb);
-        size_t xmhf_rt_code_data_size = rpb->XtVmmRuntimeDataEnd - xmhf_rt_base;
-
-        // Measure xmhf-runtime
-        printf("SL: Measure xmhf-runtime start. XMHF-runtime code and data size:0x%lX\n", xmhf_rt_code_data_size);
-        if(tpm->major == TPM12_VER_MAJOR)
-        {
-            result = sha2_256_mem_to_20bytes((void*)xmhf_rt_base, xmhf_rt_code_data_size, digest.sha1_digest);
-            if(result)
-            {
-                printf("SL: Measure xmhf-runtime with SHA1 error!\n");
-                HALT();
-            }
-        }
-        else if(tpm->major == TPM20_VER_MAJOR)
-        {
-            result = sha2_256_mem((void*)xmhf_rt_base, xmhf_rt_code_data_size, digest.sha2_256_digest);
-            if(result)
-            {
-                printf("SL: Measure xmhf-runtime with SHA256 error!\n");
-                HALT();
-            }
-        }
-        // No need to check invalid <tpm->major> again, because we have checked it.
-
-        //// Extend into PCRs
-        if(tpm->major == TPM12_VER_MAJOR)
-        {
-            hash_list_t hl;
-
-            hl.count = 1;
-            hl.entries[0].alg = TB_HALG_SHA1;
-            memcpy(&hl.entries[0].hash.sha1, digest.sha1_digest, SHA1_DIGEST_LENGTH);
-
-            result = tpm_fp->pcr_extend(tpm, 0, TPM_PCR_BOOT_STATE, &hl);
-            if(!result)
-            {
-                printf("SL (TPM1.2): Extend to PCR7 error!\n");
-                HALT();
-            }
-
-#if defined (__DRT__)
-            result = tpm_fp->pcr_extend(tpm, 2, TPM_PCR_DRTM_IMAGE, &hl);
-            if(!result)
-            {
-                printf("SL (TPM1.2): Extend to PCR17 error!\n");
-                HALT();
-            }
-#endif	//__DRT__
-        }
-        else if(tpm->major == TPM20_VER_MAJOR)
-        {
-            hash_list_t hl;
-
-            hl.count = 1;
-            hl.entries[0].alg = TB_HALG_SHA256;
-            memcpy(&hl.entries[0].hash.sha256, digest.sha2_256_digest, SHA256_DIGEST_LENGTH);
-
-            result = tpm_fp->pcr_extend(tpm, 0, TPM_PCR_BOOT_STATE, &hl);
-            if(!result)
-            {
-                printf("SL (TPM2): Extend to PCR7 error!\n");
-                HALT();
-            }
-
-#if defined (__DRT__)
-            result = tpm_fp->pcr_extend(tpm, 2, TPM_PCR_DRTM_IMAGE, &hl);
-            if(!result)
-            {
-                printf("SL (TPM2): Extend to PCR17 error!\n");
-                HALT();
-            }
-#endif	//__DRT__
-        }
-        // No need to check invalid <tpm->major> again, because we have checked it.
-
-        printf("SL: Extended xmhf-runtime measurement\n");
-    }
 
 	printf("Transferring control to runtime\n");
 	//printf("GDT=%08x, IDT=%08x, EntryPoint=%08x\n", rpb->XtVmmGdt, rpb->XtVmmIdt, rpb->XtVmmEntryPoint);
