+++ xmhf/src/xmhf-core/xmhf-secureloader/libtpm/tpm_common.h
@@ -3,6 +3,7 @@
  *
  * eXtensible, Modular Hypervisor Framework (XMHF)
  * Copyright (c) 2023 - 2024 Miao Yu
+ * Copyright (c) 2023 - 2024 Virgil Gligor
  * All Rights Reserved.
  *
  * Redistribution and use in source and binary forms, with or without
